module9-task2, Личный проект: дополнительная функциональность
